refactor: Moved data type extensions to the `util` package
