+++ app/src/main/java/pl/lambada/songsync/data/MainViewModel.kt
@@ -15,8 +15,8 @@ import pl.lambada.songsync.data.remote.lyrics_providers.spotify.SpotifyLyricsAPI
 import pl.lambada.songsync.domain.model.Release
 import pl.lambada.songsync.domain.model.Song
 import pl.lambada.songsync.domain.model.SongInfo
-import pl.lambada.songsync.data.ext.getVersion
-import pl.lambada.songsync.data.ext.toLrcFile
+import pl.lambada.songsync.util.ext.getVersion
+import pl.lambada.songsync.util.ext.toLrcFile
 import pl.lambada.songsync.ui.screens.Providers
 import java.io.FileNotFoundException
 import java.net.UnknownHostException
