+++ app/src/main/java/pl/lambada/songsync/ui/components/AppBars.kt
@@ -30,7 +30,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.navigation.NavController
 import pl.lambada.songsync.R
 import pl.lambada.songsync.domain.model.Song
-import pl.lambada.songsync.data.ext.BackPressHandler
+import pl.lambada.songsync.util.ext.BackPressHandler
 import pl.lambada.songsync.ui.Screens
 
 @OptIn(ExperimentalMaterial3Api::class)
