+++ app/src/main/java/pl/lambada/songsync/ui/screens/AboutScreen.kt
@@ -47,7 +47,7 @@ import kotlinx.coroutines.launch
 import pl.lambada.songsync.R
 import pl.lambada.songsync.data.MainViewModel
 import pl.lambada.songsync.domain.model.Release
-import pl.lambada.songsync.data.ext.getVersion
+import pl.lambada.songsync.util.ext.getVersion
 import pl.lambada.songsync.ui.components.AboutCard
 
 /**
