+++ app/src/main/java/pl/lambada/songsync/ui/screens/BrowseScreen.kt
@@ -54,7 +54,7 @@ import pl.lambada.songsync.data.MainViewModel
 import pl.lambada.songsync.data.NoTrackFoundException
 import pl.lambada.songsync.domain.model.Song
 import pl.lambada.songsync.domain.model.SongInfo
-import pl.lambada.songsync.data.ext.toLrcFile
+import pl.lambada.songsync.util.ext.toLrcFile
 import pl.lambada.songsync.ui.components.CommonTextField
 import pl.lambada.songsync.ui.components.SongCard
 import java.io.File
